fix: address PR feedback on webapp
- Rename lfm_client to lastfm_client for consistency
- Move track_counts route from lastfm.py to spotify.py (Spotify-related feature)
- Make scrobbles configurable via query params (limit, scrobbles_minimum, period, period_minimum) with defaults from config
- Move corresponding tests from test_routes_lastfm to test_routes_spotify
- Add test for scrobbles with custom query parameters
