+++ spotfm/web/routes/lastfm.py
@@ -1,4 +1,3 @@
-import csv
 from pathlib import Path
 
 from fastapi import APIRouter, Request
@@ -15,59 +14,39 @@ router = APIRouter()
 
 
 @router.get("/scrobbles", response_class=HTMLResponse)
-async def scrobbles(request: Request):
+async def scrobbles(
+    request: Request,
+    limit: str = "",
+    scrobbles_minimum: str = "",
+    period: str = "",
+    period_minimum: str = "",
+):
     redirect = await require_auth(request)
     if redirect:
         return redirect
 
     config = request.app.state.config
     lastfm_cfg = config.get("lastfm", {})
-    lfm_client = lastfm.Client(
+    lastfm_client = lastfm.Client(
         api_key=lastfm_cfg["api_key"],
         api_secret=lastfm_cfg["api_secret"],
         username=lastfm_cfg["username"],
         password_hash=lastfm_cfg["password_hash"],
     )
-    user = lastfm.User(lfm_client.client)
-    scrobbles_minimum = lastfm_cfg.get("scrobbles_minimum", 4)
-    period_minimum = lastfm_cfg.get("period_minimum")
+    user = lastfm.User(lastfm_client.client)
+
+    limit_val = int(limit) if limit else 50
+    scrobbles_min_val = int(scrobbles_minimum) if scrobbles_minimum else lastfm_cfg.get("scrobbles_minimum", 4)
+    period_val = int(period) if period else 90
+    period_min_val = int(period_minimum) if period_minimum else lastfm_cfg.get("period_minimum")
+
     # get_recent_tracks_scrobbles yields formatted strings
     tracks = list(
         user.get_recent_tracks_scrobbles(
-            limit=50,
-            scrobbles_minimum=scrobbles_minimum,
-            period=90,
-            period_minimum=period_minimum,
+            limit=limit_val,
+            scrobbles_minimum=scrobbles_min_val,
+            period=period_val,
+            period_minimum=period_min_val,
         )
     )
     return templates.TemplateResponse(request, "scrobbles.html", context={"tracks": tracks})
-
-
-@router.get("/track-counts", response_class=HTMLResponse)
-async def track_counts(request: Request):
-    redirect = await require_auth(request)
-    if redirect:
-        return redirect
-
-    config = request.app.state.config
-    log_path_raw = config.get("spotify", {}).get("track_counts_log", "")
-    if not log_path_raw:
-        return templates.TemplateResponse(
-            request, "track_counts.html", context={"rows": [], "error": "track_counts_log not configured"}
-        )
-
-    log_path = Path(log_path_raw).expanduser()
-    if not log_path.exists():
-        return templates.TemplateResponse(
-            request, "track_counts.html", context={"rows": [], "error": f"Log file not found: {log_path}"}
-        )
-
-    rows = []
-    with open(log_path, newline="") as f:
-        reader = csv.reader(f)
-        for row in reader:
-            rows.append(row)
-
-    rows.sort(key=lambda r: r[0] if r else "", reverse=True)
-
-    return templates.TemplateResponse(request, "track_counts.html", context={"rows": rows, "error": None})
